+++ FCNameColor/FCNameColor.json
@@ -5,7 +5,7 @@
   "CategoryTags": ["ui", "social"],
   "Description": "Color your FC’s tag or the entire nameplate if they are in your FC or other specified FCs.\nWorks using Lodestone data, so it can continue working inside duties as well as on different servers!",
   "InternalName": "FCNameColor",
-  "AssemblyVersion": "5.1.1.1",
+  "AssemblyVersion": "5.1.2.0",
   "RepoUrl": "https://github.com/WesselKuipers/FCNameColor",
   "IconUrl": "https://github.com/WesselKuipers/FCNameColor/raw/main/images/icon.png",
   "Tags": ["nameplate", "FC", "name", "colors"],
